requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
scipy>=1.10.0
plotly>=5.15.0